the version of alignAllen that works with my data
I'm moving more and more of my old codes to github and this is one of the function I need for my codes to work.

Some of my data is recorded with a zoom factory < 1. After shrinking the image I don't have enough pixels for the indexing using "trimH+(1:h)" in the main routine. My code fixes this.